+++ src/Concerto/TestBundle/Service/ASessionRunnerService.php
@@ -36,9 +36,9 @@ abstract class ASessionRunnerService
 
     abstract public function startNew(TestSession $session, $params, $client_ip, $client_browser, $debug = false, $max_exec_time = null);
 
-    abstract public function submit(TestSession $session, $values, $client_ip, $client_browser);
+    abstract public function submit(TestSession $session, $values, $cookies, $client_ip, $client_browser);
 
-    abstract public function backgroundWorker(TestSession $session, $values, $client_ip, $client_browser);
+    abstract public function backgroundWorker(TestSession $session, $values, $cookies, $client_ip, $client_browser);
 
     abstract public function keepAlive(TestSession $session, $client_ip, $client_browser);