+++ src/Concerto/TestBundle/Service/TestRunnerService.php
@@ -26,20 +26,19 @@ class TestRunnerService {
         return $response;
     }
 
-    public function submitToSession($session_hash, $values, $client_ip, $client_browser) {
+    public function submitToSession($session_hash, $values, $cookies, $client_ip, $client_browser) {
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_browser");
 
-
-        $response = $this->sessionService->submit($session_hash, $values, $client_ip, $client_browser);
+        $response = $this->sessionService->submit($session_hash, $values, $cookies, $client_ip, $client_browser);
         $response = json_encode($response);
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - response: $response");
         return $response;
     }
 
-    public function backgroundWorker($session_hash, $values, $client_ip, $client_browser) {
+    public function backgroundWorker($session_hash, $values, $cookies, $client_ip, $client_browser) {
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_browser");
 
-        $response = $this->sessionService->backgroundWorker($session_hash, $values, $client_ip, $client_browser);
+        $response = $this->sessionService->backgroundWorker($session_hash, $values, $cookies, $client_ip, $client_browser);
         $response = json_encode($response);
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - response: $response");
         return $response;